Started updating from new class diagram (made Rover, MarsRover, and TerrainMap)
