--- src/main/java/mars/IGeoTiff.java
@@ -1,12 +1,0 @@
-package mars;
-
-/**
- * Interface which the GeoTiff class implements
- */
-public interface IGeoTiff {
-
-    public void initTif(String fileLocation) throws Exception;
-
-    public double getValue(double x, double y) throws Exception;
-
-}
