--- src/main/java/mars/IRover.java
@@ -1,11 +1,0 @@
-package mars;
-
-/**
- * Interface which the Rover class implements
- */
-public interface IRover {
-
-    public void findOptimalPath();
-
-    public void findSuboptimalPath();
-}
