+++ src/main/java/mars/TerrainMap.java
@@ -1,0 +1,12 @@
+package mars;
+
+/**
+ * Abstract class from which all terrain maps inherit.
+ */
+public abstract class TerrainMap {
+
+    abstract void initMap(String fileLocation) throws Exception;
+
+    abstract double getValue(double x, double y) throws Exception;
+
+}
